+++ CHANGELOG.md
@@ -17,6 +17,10 @@
 * Added unique filter based on the [Jinja implementation](https://jinja.palletsprojects.com/en/stable/templates/#jinja-filters.tojson). [#16](https://github.com/gunjam/govjucks/pull/16) @gunjam
 * Added lipsum global function based on the [Jinja implementation](https://jinja.palletsprojects.com/en/stable/templates/#jinja-globals.lipsum). [#18](https://github.com/gunjam/govjucks/pull/18) @gunjam
 
+### Fixes
+
+* Macro can overwrite the caller function body in another macro. [#20](https://github.com/gunjam/govjucks/pull/20) @gunjam
+
 ## v0.1.0 (First release)
 
 This is the first release since forking nunjucks, a lot of work has been done